~~~ C O M B A T  ~  C O M M I T  ~~~
The combat commit has been commited:
And thus, Combat began in the Voclano World.

Deep in the heart of the volcano, far down
through the repo tree: The combat commit changed
the fate of the game forever.

When 2 Entities fight, the outcome will be uncertain,
each of their data will be put to the ultimate
test.  A battle of pure chance, based on some 
predefined abilities, and some of the dice roll.

Interfaces were changed, and the humans who gaze 
upon the volcano world within will see a rampage
of different things: reflections of the data
in the heart of the volcano game itself.